Restructuring to multi-module project. Improvements to Javadoc.
