+++ onshape-java/api-generator/src/main/java/com/onshape/api/generator/model/FieldMap.java
