+++ onshape-java/api-generator/src/main/resources/LICENSE
